updating the olm config with proper branding and custom resource info (#5)
Issue #, if available:
N/A

Description of changes:
Updating the OLM config file with proper CR info as well as branding. Please look at the screenshot to make sure the branding is correct for this controller. Thanks!

<img width="1422" alt="image" src="https://user-images.githubusercontent.com/88156657/213781887-7db31a92-65dd-4426-bdb9-9211efbce749.png">


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

Signed-off-by: Adam D. Cornett <adc@redhat.com>