always use ValidationMethod=DNS
Not sure when, but at some point in the past year or so, the code
generator was changed in a way that when you add fields to the `ignore`
list, that field is literally `delete`d from the associated
`aws-sdk-go/private/model/api.Shape` object. This means that you cannot
have *both* a field that is in the `ignore` list (because you don't want
it to appear in the CRD's fields) *AND* have that field referenced in
the `override_values` map for an `Operation`. Because the code that sets
an Input shape for an Operation iterates over the Input shape's
`MemberNames()` collection, and the code generator has literally
`delete`d the ignored field from the shape, meaning it cannot be
referenced in the override_values code :(

This fixes an issue in the ACM controller where the `generator.yaml`
file had the ValidationMethod field both ignored and in override_values
by adding a `sdk_create_post_build_request` hook that manually sets the
`RequestCertificateInput.ValidationMethod` field to "DNS".

Closes Issue aws-controllers-k8s/community#1701

Signed-off-by: Jay Pipes <jaypipes@gmail.com>